ci: bump v3 actions to v4
Actions based on Node 16.x.x (which many of the v3's of the official actions are)
are deprecated: https://github.blog/changelog/2023-09-22-github-actions-transitioning-from-node-16-to-node-20/

This commit bumps them to v4 to avoid any unpleasant surprises in the future.
Additionally, it folds the cache step into the Java setup step using
setup-java's cache setting.

Signed-off-by: TTtie <me@tttie.cz>